requirements.txt
kagglehub
pillow
pandas
numpy
scikit-learn
pyyaml
ultralytics
matplotlib
streamlit